+++ website/sidebars.ts
@@ -65,6 +65,7 @@ const sidebars: SidebarsConfig = {
         'reference/array',
         'reference/group',
         { type: 'doc', id: 'reference/form-node-state', label: 'useFormNodeState()' },
+        'reference/use-closest-form',
         { type: 'doc', id: 'reference/form-node-binding', label: '[formNode]' },
         'reference/form-nodes-module',
         'reference/configuration',